update all relation publish dates button

+++ app/components/super-features/super-feature-relations/super-feature-relations.js
@@ -9,15 +9,43 @@ angular.module('bulbs.cms.superFeatures.relations', [
   'bulbs.cms.titleModal',
   'bulbs.cms.utils',
   'confirmationModal',
+  'Raven',
   'statusFilter.config'
 ])
   .directive('superFeatureRelations', [
-    'CmsConfig', 'SuperFeaturesApi', 'StatusFilterOptions',
-    function (CmsConfig, SuperFeaturesApi, StatusFilterOptions) {
+    'CmsConfig', 'Raven', 'SuperFeaturesApi', 'StatusFilterOptions',
+    function (CmsConfig, Raven, SuperFeaturesApi, StatusFilterOptions) {
       return {
         controller: [
           '_', '$scope', 'Utils',
           function (_, $scope, Utils) {
+
+            var relationFormPrefix = 'relationForm_';
+            var orderingFormPrefix = 'orderingInputForm_';
+
+            $scope.wrapperForm = {};
+            $scope.makeRelationFormName = function (relation) {
+              return relationFormPrefix + relation.id;
+            };
+            $scope.makeOrderingFormName = function (relation) {
+              return orderingFormPrefix + relation.id;
+            };
+            $scope.getRelationForm = function (relation) {
+              return $scope.wrapperForm[$scope.makeRelationFormName(relation)];
+            };
+            $scope.getOrderingForm = function (relation) {
+              return $scope.wrapperForm[$scope.makeOrderingFormName(relation)];
+            };
+
+            $scope.isAtLeastOneRelationFormDirty = function () {
+              return Object.keys($scope.wrapperForm)
+                .reduce(function (isDirty, key) {
+                  if (key.startsWith(relationFormPrefix)) {
+                    return isDirty || $scope.wrapperForm[key].$dirty;
+                  }
+                  return isDirty;
+                }, false);
+            };
 
             $scope.redoOrdering = function () {
               $scope.relations.forEach(function (relation, i) {
@@ -67,9 +95,41 @@ angular.module('bulbs.cms.superFeatures.relations', [
 
             $scope.ongoingChildTransactions = {};
 
+            $scope.isAtLeastOneOngoingChildTransaction = function () {
+              return Object.keys($scope.ongoingChildTransactions)
+                .reduce(function (ongoing, relationId) {
+                  return ongoing || $scope.ongoingChildTransactions[relationId];
+                }, false);
+            };
+
+            $scope.updateChildPublishDates = function () {
+
+              if (!$scope.isAtLeastOneOngoingChildTransaction() &&
+                  !$scope.runningChildPublishDateUpdate) {
+
+                $scope.runningChildPublishDateUpdate = true;
+
+                SuperFeaturesApi
+                  .updateAllRelationPublishDates($scope.article.id)
+                  .catch(function (response) {
+                    Raven.captureMessage(
+                      'Error attempting to update all relation publish dates', {
+                        response: response
+                      }
+                    );
+                    $scope.errorMessage = 'An error occurred!';
+                  })
+                  .finally(function () {
+                    $scope.runningChildPublishDateUpdate = false;
+                  });
+              }
+            };
+
             $scope.saveChildPage = function (relation) {
 
-              if (!$scope.ongoingChildTransactions[relation.id]) {
+              if (!$scope.ongoingChildTransactions[relation.id] &&
+                  !$scope.runningChildPublishDateUpdate) {
+
                 $scope.ongoingChildTransactions[relation.id] = true;
 
                 var relationCopy = angular.copy(relation);
@@ -78,13 +138,16 @@ angular.module('bulbs.cms.superFeatures.relations', [
                 SuperFeaturesApi.updateSuperFeature(relationCopy)
                   .finally(function () {
                     $scope.ongoingChildTransactions[relation.id] = false;
+                    $scope.getRelationForm(relation).$setPristine();
                   });
               }
             };
 
             $scope.deleteChildPage = function (relation) {
 
-              if (!$scope.ongoingChildTransactions[relation.id]) {
+              if (!$scope.ongoingChildTransactions[relation.id] &&
+                  !$scope.runningChildPublishDateUpdate) {
+
                 $scope.ongoingChildTransactions[relation.id] = true;
 
                 SuperFeaturesApi.deleteSuperFeature(relation)
